+++ src/modulos/simuladorCurvas/simuladorCurvas.h
@@ -48,7 +48,7 @@ int addPointToCurve(int curveId, int tiempo, float value,aproximation_point_type
 void UpdateCurve(curve_t *curve) ;
 curve_t** newCurveArray(int size);
 void printCurves();
-int getCurveValue(int curveId);
+float getCurveValue(int curveId);
 bool validateCurve(curve_t *curve); // NUEVA: Función de validación
 void sendCurves();
 void simuladorCurvasInit(int size);
